requirements.txt
asgiref==3.4.1
click==8.0.1
cmake==3.22.1
colorama==0.4.4
Django==3.2.6
django-cors-headers==3.7.0
django-environ==0.8.1
django-extensions==3.1.3
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.22.1
opencv-python==4.5.3.56
pandas==1.3.5
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
XlsxWriter==3.0.1
